backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.108.0
uvicorn
python-multipart

# Data validation and serialization
pydantic>=2.5.2
pydantic-settings>=2.0.0

# OCR and image processing
paddleocr
paddlepaddle
opencv-python
Pillow==9.5.0

# IOPaint for advanced text inpainting and removal
iopaint==1.6.0
huggingface_hub>=0.25.0,<0.35.0

# HTTP client and async file operations
aiohttp
aiofiles
requests

# Utilities and configuration
python-dotenv
loguru

# Testing and development
pytest
pytest-asyncio
black
ruff

# File and data handling
piexif==1.1.3